Reverting QAQC changes
